+++ mobile/android/geckoview/src/main/java/org/mozilla/geckoview/WebExtensionController.java
@@ -279,31 +279,6 @@ public class WebExtensionController {
      * Called whenever an updated extension has new permissions. This is intended as an opportunity
      * for the app to prompt the user for the new permissions required by this extension.
      *
-     * @deprecated Use onUpdatePrompt(WebExtension, String[], String[], String[]) to account for
-     *     data collection permissions.
-     * @param currentlyInstalled The {@link WebExtension} that is currently installed.
-     * @param updatedExtension The {@link WebExtension} that will replace the previous extension.
-     * @param newPermissions The new permissions that are needed.
-     * @param newOrigins The new origins that are needed.
-     * @return A {@link GeckoResult} that completes to either {@link AllowOrDeny#ALLOW ALLOW} if
-     *     this extension should be update or {@link AllowOrDeny#DENY DENY} if this extension should
-     *     not be update. A null value will be interpreted as {@link AllowOrDeny#DENY DENY}.
-     */
-    @Nullable
-    @Deprecated
-    @DeprecationSchedule(id = "web-extension-on-update-prompt", version = 144)
-    default GeckoResult<AllowOrDeny> onUpdatePrompt(
-        @NonNull final WebExtension currentlyInstalled,
-        @NonNull final WebExtension updatedExtension,
-        @NonNull final String[] newPermissions,
-        @NonNull final String[] newOrigins) {
-      return null;
-    }
-
-    /**
-     * Called whenever an updated extension has new permissions. This is intended as an opportunity
-     * for the app to prompt the user for the new permissions required by this extension.
-     *
      * @param extension The {@link WebExtension} being updated.
      * @param newPermissions The new permissions that are needed.
      * @param newOrigins The new origins that are needed.
@@ -1248,34 +1223,21 @@ public class WebExtensionController {
   }
 
   private void updatePrompt(final GeckoBundle message, final EventCallback callback) {
-    final WebExtension currentExtension =
-        new WebExtension(mDelegateControllerProvider, message.getBundle("currentlyInstalled"));
-    final WebExtension updatedExtension =
-        new WebExtension(mDelegateControllerProvider, message.getBundle("updatedExtension"));
+    final WebExtension extension =
+        new WebExtension(mDelegateControllerProvider, message.getBundle("extension"));
     final String[] newPermissions = message.getStringArray("newPermissions");
     final String[] newOrigins = message.getStringArray("newOrigins");
+    final String[] newDataCollectionPermissions =
+        message.getStringArray("newDataCollectionPermissions");
 
     if (mPromptDelegate == null) {
-      Log.e(
-          LOGTAG,
-          "Tried to update extension " + currentExtension.id + " but no delegate is registered");
+      Log.e(LOGTAG, "Tried to update extension " + extension.id + " but no delegate is registered");
       return;
     }
 
-    // TODO - Bug 1974744: remove the first call when we remove the deprecated
-    // `onUpdatePrompt` method since this has been done to preserve backward
-    // compatibility.
-    GeckoResult<AllowOrDeny> promptResponse =
+    final GeckoResult<AllowOrDeny> promptResponse =
         mPromptDelegate.onUpdatePrompt(
-            currentExtension, updatedExtension, newPermissions, newOrigins);
-
-    if (promptResponse == null) {
-      final String[] newDataCollectionPermissions =
-          message.getStringArray("newDataCollectionPermissions");
-      promptResponse =
-          mPromptDelegate.onUpdatePrompt(
-              updatedExtension, newPermissions, newOrigins, newDataCollectionPermissions);
-    }
+            extension, newPermissions, newOrigins, newDataCollectionPermissions);
 
     if (promptResponse == null) {
       return;